// types/intake.ts
export type OrganizationType =
  | "For-profit"
  | "Non-profit"
  | "Not-for-profit"
  | "Government"
  | "Tribal";
export type TriState = "Yes" | "No" | "Don't Know";
// ...
export interface IntakeData {
// ...
  organizationType?: OrganizationType;
  lowIncomeOwned?: TriState; // >50% LI board/ownership
// ...
}